pyproject.toml
[project]
name = "applp"
version = "0.1.12"
description = "A library for custom GUI interface creation."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [ "corelp", "numpy", "pillow", "plotlp",]

[build-system]
requires = [ "uv_build",]
build-backend = "uv_build"

[dependency-groups]
tests = [ "pytest",]
docs = [ "sphinx", "sphinx-design", "sphinx-rtd-theme", "toml",]
anaconda = [ "ipykernel", "spyder-kernels",]

[tool.uv]
required-environments = [ "sys_platform == 'linux' and platform_machine == 'x86_64'", "sys_platform == 'win32' and (platform_machine == 'AMD64' or platform_machine == 'x86_64')",]

[tool.uv.sources.corelp]
workspace = true

[tool.uv.sources.plotlp]
workspace = true
